Update README.md
I have updated the address of the link to the page about 424 simulation on Discover website. The previous one was obsolete as "Project 424" has become "424" in the URL path of the page.
I have also updated the text to suppress any mention of "Project 424" as we call it "424" or "PERRINN 424" now.